refactor: add mocks helper and add them in random hadith
